Add OS X to some `ifdef` guards
Add OS X to some `#ifdef` guards, allowing it to use some
platform-specific code. Since OS X is similar to Linux,
most of the sections can use Linux-specific code.

There are still some sections of code guarded by ifdefs that could
possibly be enabled for OS X, for example some debugging code. However,
most of this code is not excercised in the Test compiler or JitBuilder,
so if enabled, it would not be properly tested. These sections will
be revisited when downstream projects are modified to support OS X.

Signed-off-by: Lukas Lorimer <lukasl@ca.ibm.com>