Changing default behavior for compilation aggressiveness level 3(AGGRESSIVE_AOT)
The JIT compiler has several heuristic "policies" referred to as
aggressivenessLevels. The current change-set fine-tunes the behavior of
compilation aggressivenessLevel 3 (AGGRESSIVE_AOT) to deliver a better
balance between throughput and CPU consumption.
Specifically, two changes are performed under the hood:
(1) The GRACostBenefitModel feature is disabled. This is a heuristic that
excludes some candidates from the global register allocator optimization,
based on estimates regarding CPU overhead and potential benefits.
The reason is that this feature introduced too much of a throughput loss
compared to the CPU savings it provided (we strive for a 1:10 ratio)
(2) To compensate for the CPU increase induced by the first change we reduced
one of the inlining thresholds: bigCalleeThreshold. This represents the
maximum size of a method (measured in bytecodes) that will be considered for
inlining. The net effect of these two changes combined is that throughput
improved by ~1.5% and CPU overhead reduced by 5% in the benchmarks we
considered.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>