Refactor isPotentialOSRPoint query in two
The isPotentialOSRPoint query is used in two way. The first mode of
operation is the ILGen or other related infrastructure asking if a given
node could be an OSR point. This is a property of the node itself and
does not rely on the OSR infrastructure determining if OSR is actually
supported at that point - it is a conservative yes we may need to
support OSR at this point rather than a definite we can OSR at this
point. The second mode of operation is with a TreeTop where the
optimizer needs to know if OSR can be performed at a particular point.
This question requires that the first mode of operation says yes we may
OSR as well as the OSR infrastructure saying yes OSR is supported.

This change breaks apart the these two uses - isPotentialOSRPoint
answers the first query while isPotentialOSRPointWithSupport answers the
second query.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>